requirements.txt
PyMuPDF
google-generativeai
python-dotenv
flask